+++ src/BoardRender.js
@@ -1,66 +1,4 @@
-
-let gradient_1 = [
-    {weight: 0,
-        r: 255, g: 255, b: 255,},
-    {weight: 20,
-        r: 50, g: 250, b: 150,},
-    {weight: 80,
-        r: 50, g: 150, b: 250,},
-    {weight: 100,
-        r: 0, g: 0, b: 0,},
-];
-let gradient_2 = [
-    {weight: 0,
-        r: 255, g: 0, b: 40,},
-    {weight: 50,
-        r: 255, g: 255, b: 255,},
-    {weight: 100,
-        r: 0, g: 40, b: 255,},
-];
-let gradient_3 = [
-    {weight: 0,
-        r: 255, g: 255, b: 200,},
-    {weight: 50,
-        r: 100, g: 200, b: 150,},
-    {weight: 100,
-        r: 60, g: 52, b: 45,},
-];
-let gradient_4 = [
-    {weight: 0,
-        r: 255, g: 240, b: 250,},
-    {weight: 25,
-        r: 50, g: 255, b: 40,},
-    {weight: 50,
-        r: 120, g: 125, b: 135,},
-    {weight: 75,
-        r: 20, g: 20, b: 200,},
-    {weight: 100,
-        r: 0, g: 0, b: 0,},
-];
-
-//note: add values to lookup table so they dont need to be recalculated
-let gradientPointValue = (gradient, weight) => {
-    let i;
-    //find colors surounding weight
-    for(i = 0; i < gradient.length; i++){
-        if(gradient[i].weight > weight) break;
-    }
-    let c1 = gradient[i];
-    let c0 = gradient[i - 1];
-    //console.log(c1);
-    //console.log(c0);
-
-    let npw = weight - c0.weight; // normalize point weight
-    let ncw = c1.weight - c0.weight // normalize color weight
-
-    // normalize weight, multiply by slope, add to first color
-    let r = Math.round(c0.r + npw * (c1.r - c0.r) / ncw); 
-    let g = Math.round(c0.g + npw * (c1.g - c0.g) / ncw); 
-    let b = Math.round(c0.b + npw * (c1.b - c0.b) / ncw); 
-
-    return `rgb(${r}, ${g}, ${b})`;
-}
-
+import {gradientPointValueFromChoice} from './ColorMap.js';
 /**
  * events:
  * gameWon
@@ -103,16 +41,13 @@ export default class BoardRender extends EventTarget{
 
     generateGradientMap(){
 
-        let set = [gradient_1, gradient_2, gradient_3, gradient_4];
-        let choice = Math.floor(Math.random() * set.length);
-        let gradient = set[choice];
-
+        let choice = Math.floor(Math.random() * 4); //4 gradients to choose from, coupled code, TODO: refactor
         console.log("using color gradient: " + (choice + 1));
         
 
         this.colormap = [];
         for(let i = 0; i < 100; i++){
-            this.colormap[i] = gradientPointValue(gradient, i);
+            this.colormap[i] = gradientPointValueFromChoice(choice, i);
         }
     }
     destroy(){
@@ -198,42 +133,19 @@ export default class BoardRender extends EventTarget{
 
         //map value => color value
         let kWeight = 2;
-        let normVal = Math.round(100* (-1 * targetData.value + kWeight/2) / kWeight );
+        let normVal = Math.round(100* (1 * targetData.value + kWeight/2) / kWeight );
         let cappedVal = Math.min( 99 , Math.max (0 , normVal ));
         targetElement.style.background = this.colormap[cappedVal];
         
-        /*
-        //assumes .value is in the range [-1,1]
-        //colorval mapping is slapdashed as fuck but need MVP
-        let normalize_midpoint = 0.6;
-        let normalize_weight = 1.3;
-
-        //manually cap color val at 0 and 255
-        let colorVal =  Math.max(0, Math.min((targetData.value/normalize_weight + normalize_midpoint) * 255, 255));
-
-        
-        switch(this.color){
-            case 0:
-                targetElement.style.background = `rgb(${Math.floor(colorVal / 1)},${Math.floor(colorVal / 2)},${Math.floor(colorVal / 1.2)})`;
-                //targetElement.style.background = gradientPointValue(gradient_1, cappedVal);
-                break;
-            case 1:
-                targetElement.style.background = `rgb(${Math.floor(colorVal / 1.2)},${Math.floor(colorVal / 1)},${Math.floor(colorVal / 2)})`;
-                //targetElement.style.background = gradientPointValue(gradient_2, cappedVal);
-                break;
-            default:
-                targetElement.style.background = `rgb(${Math.floor(colorVal / 2)},${Math.floor(colorVal / 1.2)},${Math.floor(colorVal / 1)})`;
-                //targetElement.style.background = gradientPointValue(gradient_3, cappedVal);
-        }
-        */
-        
-
         //color mines
         if(targetData.isMine){
-            if(targetData.value >= 0){
-                targetElement.style.background = this.colormap[0];
-            }else{
+            if(targetData.value > 0){
                 targetElement.style.background = this.colormap[100];
+            }else if(targetData.value < 0){
+                targetElement.style.background = this.colormap[0];
+            }
+            else{
+                targetElement.style.background = this.colormap[50]
             }
             
         }